+++ R/bootstrap_model.R
@@ -4,7 +4,7 @@
 #'
 #' @param model Statistical model.
 #' @param iterations The number of draws to simulate/bootstrap.
-#' @param verbose Hide possible refit messages.
+#' @param verbose Show or hide possible warnings and messages.
 #' @param ... Arguments passed to or from other methods.
 #'
 #' @return A data frame.