+++ R/simulate_parameters.R
@@ -70,6 +70,7 @@ simulate_parameters.default <- function(model, iterations = 1000, centrality = "
   class(out) <- c("parameters_simulate", "see_parameters_simulate", class(out))
   attr(out, "object_name") <- deparse(substitute(model), width.cutoff = 500)
   attr(out, "iterations") <- iterations
+  attr(out, "ci") <- ci
 
   out
 }
@@ -90,6 +91,32 @@ simulate_parameters.multinom <- function(model, iterations = 1000, centrality = 
   class(out) <- c("parameters_simulate", "see_parameters_simulate", class(out))
   attr(out, "object_name") <- deparse(substitute(model), width.cutoff = 500)
   attr(out, "iterations") <- iterations
+  attr(out, "ci") <- ci
+
+  out
+}
+
+
+
+#' @rdname simulate_parameters
+#' @export
+simulate_parameters.glmmTMB <- function(model, iterations = 1000, centrality = "median", ci = .95, ci_method = "quantile", test = "p-value", ...) {
+  data <- simulate_model(model, iterations = iterations, ...)
+  out <- .summary_bootstrap(data = data, test = test, centrality = centrality, ci = ci, ci_method = ci_method, ...)
+
+  params <- insight::get_parameters(model, ...)
+  if ("Effects" %in% colnames(params) && .n_unique(params$Effects) > 1) {
+    out$Effects <- params$Effects
+  }
+
+  if ("Component" %in% colnames(params) && .n_unique(params$Component) > 1) {
+    out$Component <- params$Component
+  }
+
+  class(out) <- c("parameters_simulate", "see_parameters_simulate", class(out))
+  attr(out, "object_name") <- deparse(substitute(model), width.cutoff = 500)
+  attr(out, "iterations") <- iterations
+  attr(out, "ci") <- ci
 
   out
 }